added test for index and remove route in recipe admin
