--- tests/Controller/RecipeControllerTest.php
@@ -1,117 +1,0 @@
-<?php
-
-namespace App\Test\Controller;
-
-use App\Entity\Recipe;
-use App\Repository\RecipeRepository;
-use Symfony\Bundle\FrameworkBundle\KernelBrowser;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-
-class RecipeControllerTest extends WebTestCase
-{
-    private KernelBrowser $client;
-    private RecipeRepository $repository;
-    private string $path = '/admin/recipe/';
-
-    protected function setUp(): void
-    {
-        $this->client = static::createClient();
-        $this->repository = static::getContainer()->get('doctrine')->getRepository(Recipe::class);
-
-        foreach ($this->repository->findAll() as $object) {
-            $this->repository->remove($object, true);
-        }
-    }
-
-    public function testIndex(): void
-    {
-        $crawler = $this->client->request('GET', $this->path);
-
-        self::assertResponseStatusCodeSame(200);
-        self::assertPageTitleContains('Recipe index');
-
-        // Use the $crawler to perform additional assertions e.g.
-        // self::assertSame('Some text on the page', $crawler->filter('.p')->first());
-    }
-
-    public function testNew(): void
-    {
-        $originalNumObjectsInRepository = count($this->repository->findAll());
-
-        $this->markTestIncomplete();
-        $this->client->request('GET', sprintf('%snew', $this->path));
-
-        self::assertResponseStatusCodeSame(200);
-
-        $this->client->submitForm('Save', [
-            'recipe[title]' => 'Testing',
-            'recipe[user]' => 'Testing',
-        ]);
-
-        self::assertResponseRedirects('/admin/recipe/');
-
-        self::assertSame($originalNumObjectsInRepository + 1, count($this->repository->findAll()));
-    }
-
-    public function testShow(): void
-    {
-        $this->markTestIncomplete();
-        $fixture = new Recipe();
-        $fixture->setTitle('My Title');
-        $fixture->setUser('My Title');
-
-        $this->repository->save($fixture, true);
-
-        $this->client->request('GET', sprintf('%s%s', $this->path, $fixture->getId()));
-
-        self::assertResponseStatusCodeSame(200);
-        self::assertPageTitleContains('Recipe');
-
-        // Use assertions to check that the properties are properly displayed.
-    }
-
-    public function testEdit(): void
-    {
-        $this->markTestIncomplete();
-        $fixture = new Recipe();
-        $fixture->setTitle('My Title');
-        $fixture->setUser('My Title');
-
-        $this->repository->save($fixture, true);
-
-        $this->client->request('GET', sprintf('%s%s/edit', $this->path, $fixture->getId()));
-
-        $this->client->submitForm('Update', [
-            'recipe[title]' => 'Something New',
-            'recipe[user]' => 'Something New',
-        ]);
-
-        self::assertResponseRedirects('/admin/recipe/');
-
-        $fixture = $this->repository->findAll();
-
-        self::assertSame('Something New', $fixture[0]->getTitle());
-        self::assertSame('Something New', $fixture[0]->getUser());
-    }
-
-    public function testRemove(): void
-    {
-        $this->markTestIncomplete();
-
-        $originalNumObjectsInRepository = count($this->repository->findAll());
-
-        $fixture = new Recipe();
-        $fixture->setTitle('My Title');
-        $fixture->setUser('My Title');
-
-        $this->repository->save($fixture, true);
-
-        self::assertSame($originalNumObjectsInRepository + 1, count($this->repository->findAll()));
-
-        $this->client->request('GET', sprintf('%s%s', $this->path, $fixture->getId()));
-        $this->client->submitForm('Delete');
-
-        self::assertSame($originalNumObjectsInRepository, count($this->repository->findAll()));
-        self::assertResponseRedirects('/admin/recipe/');
-    }
-}
